String constants representing item's name refactoring
