+++ src/main/java/net/gtemgoua/kata/gildedrose/domain/Item.java
@@ -1,4 +1,4 @@
-package net.gtemgoua.kata.gildedrose;
+package net.gtemgoua.kata.gildedrose.domain;
 
 public class Item {
