+++ src/test/java/net/gtemgoua/kata/gildedrose/GildedRoseTest.java
@@ -4,6 +4,8 @@ import static org.junit.Assert.assertEquals;
 
 import org.junit.Test;
 
+import net.gtemgoua.kata.gildedrose.domain.Item;
+
 public class GildedRoseTest {
 
     @Test
